+++ ViewHoldingsRecord.js
@@ -12,16 +12,17 @@ import { Row, Col } from '@folio/stripes-components/lib/LayoutGrid';
 import Headline from '@folio/stripes-components/lib/Headline';
 import IconButton from '@folio/stripes-components/lib/IconButton';
 import AppIcon from '@folio/stripes-components/lib/AppIcon';
+import ViewMetaData from '@folio/stripes-smart-components/lib/ViewMetaData';
 
 import craftLayerUrl from '@folio/stripes-components/util/craftLayerUrl';
 
 import HoldingsForm from './edit/holdings/HoldingsForm';
-import ViewMetadata from './ViewMetadata';
 
 class ViewHoldingsRecord extends React.Component {
   static manifest = Object.freeze({
     query: {},
-    locationQuery: {},
+    permanentLocationQuery: {},
+    temporaryLocationQuery: {},
     holdingsRecords: {
       type: 'okapi',
       path: 'holdings-storage/holdings/:{holdingsrecordid}',
@@ -33,29 +34,21 @@ class ViewHoldingsRecord extends React.Component {
       type: 'okapi',
       path: 'instance-storage/instances/:{id}',
     },
-    locations: {
+    permanentLocation: {
       type: 'okapi',
-      path: 'locations/%{locationQuery.id}',
+      path: 'locations/%{permanentLocationQuery.id}',
     },
+    temporaryLocation: {
+      type: 'okapi',
+      path: 'locations/%{temporaryLocationQuery.id}',
+    },
+
     platforms: {
       type: 'okapi',
       path: 'platforms',
       records: 'platforms',
     },
   });
-
-  static getDerivedStateFromProps(nextProps) {
-    const { resources } = nextProps;
-    const holdingsRecords = (resources.holdingsRecords || {}).records || [];
-    const locationQuery = resources.locationQuery;
-    const holding = holdingsRecords[0];
-
-    if (holding && (!locationQuery.id || locationQuery.id !== holding.permanentLocationId)) {
-      nextProps.mutator.locationQuery.update({ id: holding.permanentLocationId });
-    }
-
-    return null;
-  }
 
   constructor(props) {
     super(props);
@@ -65,7 +58,24 @@ class ViewHoldingsRecord extends React.Component {
       },
     };
     this.craftLayerUrl = craftLayerUrl.bind(this);
-    this.cViewMetadata = props.stripes.connect(ViewMetadata);
+    this.cViewMetaData = props.stripes.connect(ViewMetaData);
+  }
+
+  static getDerivedStateFromProps(nextProps) {
+    const { resources } = nextProps;
+    const holdingsRecords = (resources.holdingsRecords || {}).records || [];
+    const permanentLocationQuery = resources.permanentLocationQuery;
+    const temporaryLocationQuery = resources.temporaryLocationQuery;
+    const holding = holdingsRecords[0];
+
+    if (holding && (!permanentLocationQuery.id || permanentLocationQuery.id !== holding.permanentLocationId)) {
+      nextProps.mutator.permanentLocationQuery.update({ id: holding.permanentLocationId });
+    }
+    if (holding && holding.temporaryLocationId && (!temporaryLocationQuery.id || temporaryLocationQuery.id !== holding.temporaryLocationId)) {
+      nextProps.mutator.temporaryLocationQuery.update({ id: holding.temporaryLocationId });
+    }
+
+    return null;
   }
 
   // Edit Holdings records handlers
@@ -119,16 +129,17 @@ class ViewHoldingsRecord extends React.Component {
   }
 
   render() {
-    const { location, resources: { holdingsRecords, instances1, platforms, locations }, referenceTables, okapi } = this.props;
+    const { location, resources: { holdingsRecords, instances1, platforms, permanentLocation, temporaryLocation }, referenceTables, okapi } = this.props;
 
     if (!holdingsRecords || !holdingsRecords.hasLoaded
         || !instances1 || !instances1.hasLoaded
-        || !locations || !locations.hasLoaded
+        || !permanentLocation || !permanentLocation.hasLoaded
         || !platforms || !platforms.hasLoaded) return <div>Awaiting resources</div>;
 
     const holdingsRecord = holdingsRecords.records[0];
     const instance = instances1.records[0];
-    const holdingLocation = locations.records[0];
+    const holdingsPermanentLocation = permanentLocation.records[0];
+    const holdingsTemporaryLocation = holdingsRecord.temporaryLocationId ? temporaryLocation.records[0] : null;
 
     referenceTables.platforms = platforms.records;
 
@@ -162,7 +173,7 @@ class ViewHoldingsRecord extends React.Component {
             defaultWidth={this.props.paneWidth}
             paneTitle={
               <div style={{ textAlign: 'center' }}>
-                <AppIcon app="inventory" iconKey="holdings" size="small" />&nbsp;<strong>{holdingsRecord.permanentLocationId ? `${holdingLocation.name} >` : null} {_.get(holdingsRecord, ['callNumber'], '')}</strong>&nbsp;
+                <AppIcon app="inventory" iconKey="holdings" size="small" />&nbsp;<strong>{holdingsRecord.permanentLocationId ? `${holdingsPermanentLocation.name} >` : null} {_.get(holdingsRecord, ['callNumber'], '')}</strong>&nbsp;
                 <div>
                   {formatMsg({ id: 'ui-inventory.holdings' })}
                 </div>
@@ -180,51 +191,74 @@ class ViewHoldingsRecord extends React.Component {
                 }
               </Col>
             </Row>
-
+            <Row>
+              <Col sm={12}>
+                <AppIcon app="inventory" iconKey="holdings" size="small" /> {formatMsg({ id: 'ui-inventory.holdingsRecord' })}
+              </Col>
+            </Row>
+            <Row>
+              <Col sm={12}>
+                <Headline size="medium" margin="medium">
+                  {holdingsRecord.permanentLocationId ? holdingsPermanentLocation.name : null} &gt; {_.get(holdingsRecord, ['callNumber'], '')}
+                </Headline>
+              </Col>
+            </Row>
             <Accordion
               open={this.state.accordions.holdingsAccordion}
               id="holdingsAccordion"
               onToggle={this.handleAccordionToggle}
               label={formatMsg({ id: 'ui-inventory.holdingsData' })}
             >
-              <Row>
-                <Col sm={12}>
-                  <AppIcon app="inventory" iconKey="holdings" size="small" /> {formatMsg({ id: 'ui-inventory.holdingsRecord' })}
-                </Col>
-              </Row>
               <br />
               { (holdingsRecord.metadata && holdingsRecord.metadata.createdDate) &&
-                <this.cViewMetadata metadata={holdingsRecord.metadata} />
+                <this.cViewMetaData metadata={holdingsRecord.metadata} />
               }
               <Row>
-                <Col sm={12}>
-                  <Headline size="medium" margin="medium">
-                    {holdingsRecord.permanentLocationId ? holdingLocation.name : null} &gt; {_.get(holdingsRecord, ['callNumber'], '')}
-                  </Headline>
-                </Col>
-              </Row>
-              <Row>
-                <Col smOffset={1} sm={4}>
+                <Col smOffset={0} sm={4}>
                   <KeyValue label={formatMsg({ id: 'ui-inventory.holdingsId' })} value={_.get(holdingsRecord, ['id'], '')} />
                 </Col>
               </Row>
+              <Accordion
+                open={this.state.accordions.locationAccordion}
+                id="locationAccordion"
+                onToggle={this.handleAccordionToggle}
+                label={formatMsg({ id: 'ui-inventory.location' })}
+              >
+                <Row>
+                  <Col smOffset={0} sm={4}>
+                    <strong>{formatMsg({ id: 'ui-inventory.holdingsLocation' })}</strong>
+                  </Col>
+                </Row>
+                <br />
+                { ((holdingsRecord.permanentLocationId) || (holdingsRecord.temporaryLocationId)) &&
+                  <Row>
+                    <Col smOffset={0} sm={4}>
+                      <KeyValue label={formatMsg({ id: 'ui-inventory.permanent' })} value={holdingsPermanentLocation.name} />
+                    </Col>
+                    <Col>
+                      <KeyValue label={formatMsg({ id: 'ui-inventory.temporary' })} value={holdingsTemporaryLocation ? holdingsTemporaryLocation.name : null} />
+                    </Col>
+                  </Row>
+
+                }
+              </Accordion>
               { (holdingsRecord.electronicLocation && holdingsRecord.electronicLocation.platformId) &&
                 <Row>
-                  <Col smOffset={1} sm={4}>
+                  <Col smOffset={0} sm={4}>
                     <KeyValue label={formatMsg({ id: 'ui-inventory.platform' })} value={_.get(holdingsRecord, ['electronicLocation', 'platformId'], '') ? platforms.records.find(platform => _.get(holdingsRecord, ['electronicLocation', 'platformId']) === platform.id).name : null} />
                   </Col>
                 </Row>
               }
               { (holdingsRecord.electronicLocation && holdingsRecord.electronicLocation.uri) &&
                 <Row>
-                  <Col smOffset={1} sm={4}>
+                  <Col smOffset={0} sm={4}>
                     <KeyValue label={formatMsg({ id: 'ui-inventory.uri' })} value={_.get(holdingsRecord, ['electronicLocation', 'uri'], '')} />
                   </Col>
                 </Row>
               }
               { (holdingsRecord.holdingsStatements.length > 0) &&
                 <Row>
-                  <Col smOffset={1} sm={4}>
+                  <Col smOffset={0} sm={4}>
                     <KeyValue label={formatMsg({ id: 'ui-inventory.holdingsStatements' })} value={_.get(holdingsRecord, ['holdingsStatements'], []).map((line, i) => <div key={i}>{line}</div>)} />
                   </Col>
                 </Row>
@@ -277,6 +311,12 @@ ViewHoldingsRecord.propTypes = {
     locations: PropTypes.shape({
       records: PropTypes.arrayOf(PropTypes.object),
     }),
+    permanentLocation: PropTypes.shape({
+      records: PropTypes.arrayOf(PropTypes.object),
+    }),
+    temporaryLocation: PropTypes.shape({
+      records: PropTypes.arrayOf(PropTypes.object),
+    }),
     platforms: PropTypes.shape({
       records: PropTypes.arrayOf(PropTypes.object),
     }),
@@ -291,7 +331,8 @@ ViewHoldingsRecord.propTypes = {
       POST: PropTypes.func.isRequired,
     }),
     query: PropTypes.object.isRequired,
-    locationQuery: PropTypes.object.isRequired,
+    permanentLocationQuery: PropTypes.object.isRequired,
+    temporaryLocationQuery: PropTypes.object.isRequired,
   }),
   onCloseViewHoldingsRecord: PropTypes.func.isRequired,
 };