+++ ViewInstance.js
@@ -6,6 +6,7 @@ import queryString from 'query-string';
 import { graphql } from 'react-apollo';
 import gql from 'graphql-tag';
 
+import TitleManager from '@folio/stripes-core/src/components/TitleManager';
 import Pane from '@folio/stripes-components/lib/Pane';
 import PaneMenu from '@folio/stripes-components/lib/PaneMenu';
 import { Accordion, ExpandAllButton } from '@folio/stripes-components/lib/Accordion';
@@ -17,6 +18,7 @@ import IconButton from '@folio/stripes-components/lib/IconButton';
 import AppIcon from '@folio/stripes-components/lib/AppIcon';
 import Icon from '@folio/stripes-components/lib/Icon';
 import Headline from '@folio/stripes-components/lib/Headline';
+import ViewMetaData from '@folio/stripes-smart-components/lib/ViewMetaData';
 
 import craftLayerUrl from '@folio/stripes-components/util/craftLayerUrl';
 
@@ -27,7 +29,6 @@ import InstanceForm from './edit/InstanceForm';
 import HoldingsForm from './edit/holdings/HoldingsForm';
 import ViewHoldingsRecord from './ViewHoldingsRecord';
 import ViewItem from './ViewItem';
-import ViewMetadata from './ViewMetadata';
 import makeConnectedInstance from './ConnectedInstance';
 
 const emptyObj = {};
@@ -127,7 +128,7 @@ class ViewInstance extends React.Component {
     this.cHoldings = this.props.stripes.connect(Holdings);
     this.cViewHoldingsRecord = this.props.stripes.connect(ViewHoldingsRecord);
     this.cViewItem = this.props.stripes.connect(ViewItem);
-    this.cViewMetadata = this.props.stripes.connect(ViewMetadata);
+    this.cViewMetaData = this.props.stripes.connect(ViewMetaData);
 
     this.craftLayerUrl = craftLayerUrl.bind(this);
   }
@@ -267,6 +268,7 @@ class ViewInstance extends React.Component {
         dismissible
         onClose={this.props.onClose}
       >
+        <TitleManager record={instance.title} />
         <Row end="xs"><Col xs><ExpandAllButton accordionStatus={this.state.accordions} onToggle={this.handleExpandAll} /></Col></Row>
         <Accordion
           open={this.state.accordions.instanceAccordion}
@@ -275,7 +277,7 @@ class ViewInstance extends React.Component {
           label={formatMsg({ id: 'ui-inventory.instanceData' })}
         >
           { (instance.metadata && instance.metadata.createdDate) &&
-            <this.cViewMetadata metadata={instance.metadata} />
+            <this.cViewMetaData metadata={instance.metadata} />
           }
           <Row>
             <Col xs={12}>