+++ ViewItem.js
@@ -13,11 +13,11 @@ import { Row, Col } from '@folio/stripes-components/lib/LayoutGrid';
 import Headline from '@folio/stripes-components/lib/Headline';
 import IconButton from '@folio/stripes-components/lib/IconButton';
 import AppIcon from '@folio/stripes-components/lib/AppIcon';
+import ViewMetaData from '@folio/stripes-smart-components/lib/ViewMetaData';
 
 import craftLayerUrl from '@folio/stripes-components/util/craftLayerUrl';
 
 import ItemForm from './edit/items/ItemForm';
-import ViewMetadata from './ViewMetadata';
 
 class ViewItem extends React.Component {
   static manifest = Object.freeze({
@@ -66,6 +66,21 @@ class ViewItem extends React.Component {
       records: 'users',
     },
   });
+
+  constructor(props) {
+    super(props);
+    this.state = {
+      accordions: {
+        itemAccordion: true,
+      },
+      loan: null,
+      borrower: null,
+      loanStatusDate: null,
+    };
+
+    this.craftLayerUrl = craftLayerUrl.bind(this);
+    this.cViewMetaData = props.stripes.connect(ViewMetaData);
+  }
 
   /**
    * If a loan is retrieved matching this item, retrieve the corresponding
@@ -134,21 +149,6 @@ class ViewItem extends React.Component {
     return null;
   }
 
-  constructor(props) {
-    super(props);
-    this.state = {
-      accordions: {
-        itemAccordion: true,
-      },
-      loan: null,
-      borrower: null,
-      loanStatusDate: null,
-    };
-
-    this.craftLayerUrl = craftLayerUrl.bind(this);
-    this.cViewMetadata = props.stripes.connect(ViewMetadata);
-  }
-
   onClickEditItem = (e) => {
     if (e) e.preventDefault();
     this.props.mutator.query.update({ layer: 'editItem' });
@@ -213,7 +213,9 @@ class ViewItem extends React.Component {
     const item = items.records[0];
     const holdingsRecord = holdingsRecords.records[0];
     const { locationsById } = referenceTables;
-    const holdingLocation = locationsById[holdingsRecord.permanentLocationId];
+    const permanentHoldingsLocation = locationsById[holdingsRecord.permanentLocationId];
+    const temporaryHoldingsLocation = locationsById[holdingsRecord.temporaryLocationId];
+
     const requestRecords = (requests || {}).records || [];
     const query = location.search ? queryString.parse(location.search) : {};
 
@@ -236,7 +238,7 @@ class ViewItem extends React.Component {
       </PaneMenu>
     );
 
-    const labelLocation = _.get(holdingLocation, ['name'], '');
+    const labelPermanentHoldingsLocation = _.get(permanentHoldingsLocation, ['name'], '');
     const labelCallNumber = holdingsRecord.callNumber || '';
 
     let requestLink = 0;
@@ -280,7 +282,7 @@ class ViewItem extends React.Component {
                 <span><em>, </em><em>{instance.publication[0].publisher}{instance.publication[0].dateOfPublication ? `, ${instance.publication[0].dateOfPublication}` : ''}</em></span>
                 }
                 <div>
-                  { `${formatMsg({ id: 'ui-inventory.holdingsColon' })} ${labelLocation} > ${labelCallNumber}`}
+                  { `${formatMsg({ id: 'ui-inventory.holdingsColon' })} ${labelPermanentHoldingsLocation} > ${labelCallNumber}`}
                 </div>
               </Col>
             </Row>
@@ -297,7 +299,7 @@ class ViewItem extends React.Component {
               </Row>
               <br />
               { (item.metadata && item.metadata.createdDate) &&
-                <this.cViewMetadata metadata={item.metadata} />
+                <this.cViewMetaData metadata={item.metadata} />
               }
               { (item.barcode) &&
                 <Row>
@@ -314,13 +316,6 @@ class ViewItem extends React.Component {
                   <KeyValue label={formatMsg({ id: 'ui-inventory.itemId' })} value={_.get(item, ['id'], '')} />
                 </Col>
               </Row>
-              { (item.temporaryLocation) &&
-                <Row>
-                  <Col smOffset={0} sm={4}>
-                    <KeyValue label={formatMsg({ id: 'ui-inventory.temporaryLocation' })} value={_.get(item, ['temporaryLocation', 'name'], '')} />
-                  </Col>
-                </Row>
-              }
               { (item.permanentLoanType) &&
                 <Row>
                   <Col smOffset={0} sm={4}>
@@ -397,6 +392,53 @@ class ViewItem extends React.Component {
                 </Col>
                 <Col smOffset={0} sm={4}>
                   <KeyValue label={intl.formatMessage({ id: 'ui-inventory.item.availability.dueDate' })} value={this.state.loan ? formatDateTime(this.state.loan.dueDate) : '-'} />
+                </Col>
+              </Row>
+            </Accordion>
+            <Accordion
+              open={this.state.accordions.locationAccordion}
+              id="locationAccordion"
+              onToggle={this.handleAccordionToggle}
+              label={formatMsg({ id: 'ui-inventory.location' })}
+            >
+              <Row>
+                <Col smOffset={0} sm={4}>
+                  <strong>{formatMsg({ id: 'ui-inventory.holdingsLocation' })}</strong>
+                </Col>
+              </Row>
+              <br />
+              <Row>
+                <Col smOffset={0} sm={4}>
+                  <KeyValue label={formatMsg({ id: 'ui-inventory.permanentLocation' })} value={_.get(permanentHoldingsLocation, ['name'], '')} />
+                </Col>
+                <Col sm={4}>
+                  <KeyValue label={formatMsg({ id: 'ui-inventory.temporaryLocation' })} value={_.get(temporaryHoldingsLocation, ['name'], '')} />
+                </Col>
+              </Row>
+              <Row>
+                <Col smOffset={0} sm={4}>
+                  <strong>{formatMsg({ id: 'ui-inventory.itemLocation' })}</strong>
+                </Col>
+              </Row>
+              <br />
+              <Row>
+                <Col smOffset={0} sm={4}>
+                  <KeyValue label={formatMsg({ id: 'ui-inventory.permanentLocation' })} value={_.get(item, ['permanentLocation', 'name'], 'Inherit from holdings')} />
+                </Col>
+                <Col sm={4}>
+                  <KeyValue label={formatMsg({ id: 'ui-inventory.temporaryLocation' })} value={_.get(item, ['temporaryLocation', 'name'], 'Inherit from holdings')} />
+                </Col>
+              </Row>
+              <br />
+              <Row>
+                <Col smOffset={0} sm={4}>
+                  <strong>{formatMsg({ id: 'ui-inventory.effectiveLocation' })}</strong>
+                </Col>
+              </Row>
+              <br />
+              <Row>
+                <Col smOffset={0} sm={4}>
+                  {_.get(item, ['effectiveLocation', 'name'], '')}
                 </Col>
               </Row>
             </Accordion>