+++ edit/holdings/HoldingsForm.js
@@ -32,6 +32,7 @@ class HoldingsForm extends React.Component {
     submitting: PropTypes.bool,
     copy: PropTypes.bool,
     onCancel: PropTypes.func,
+    onSubmit: PropTypes.func,
     initialValues: PropTypes.object,
     instance: PropTypes.object,
     referenceTables: PropTypes.object.isRequired,
@@ -42,35 +43,79 @@ class HoldingsForm extends React.Component {
   constructor() {
     super();
     this.state = {
-      confirmLocation: false,
+      confirmPermanentLocation: false,
+      confirmTemporaryLocation: false,
     };
   }
 
   componentDidMount() {
     const { initialValues } = this.props;
-    const prevLocation = initialValues.temporaryLocation || {};
+    const prevPermanentLocation = initialValues.permanentLocation || {};
     // eslint-disable-next-line react/no-did-mount-set-state
-    this.setState({ prevLocation });
+    this.setState({ prevPermanentLocation });
+
+    const prevTemporaryLocation = initialValues.temporaryLocation || {};
+    // eslint-disable-next-line react/no-did-mount-set-state
+    this.setState({ prevTemporaryLocation });
+
+    this.onSave = this.onSave.bind(this);
   }
 
-  selectLocation(location) {
-    if (!location.id) return;
+  onSave(data) {
+    if (!data.temporaryLocationId) {
+      delete data.temporaryLocationId;
+    }
+    if (!data.permanentLocationId) {
+      delete data.permanentLocationId;
+    }
 
-    if (location.isActive) {
-      this.setState({ prevLocation: location });
-      setTimeout(() => this.props.change('permanentLocationId', location.id));
+    this.props.onSubmit(data);
+  }
+
+  selectPermanentLocation(permanentLocation) {
+    if (!permanentLocation) {
+      this.props.change('permanentLocationId', '');
+      return;
+    }
+
+    if (permanentLocation.isActive) {
+      this.setState({ prevPermanentLocation: permanentLocation });
+      setTimeout(() => this.props.change('permanentLocationId', permanentLocation.id));
     } else {
-      this.setState({ confirmLocation: true, location });
+      this.setState({ confirmPermanentLocation: true, permanentLocation });
     }
   }
 
-  confirmLocation(confirm) {
-    const { location, prevLocation } = this.state;
-    const confirmLocation = false;
-    const value = (confirm) ? location.id : prevLocation.id;
-    const prevLoc = (confirm) ? location : prevLocation;
+  selectTemporaryLocation(temporaryLocation) {
+    if (!temporaryLocation) {
+      this.props.change('temporaryLocationId', '');
+      return;
+    }
+
+    if (temporaryLocation.isActive) {
+      this.setState({ prevTemporaryLocation: temporaryLocation });
+      setTimeout(() => this.props.change('temporaryLocationId', temporaryLocation.id));
+    } else {
+      this.setState({ confirmTemporaryLocation: true, temporaryLocation });
+    }
+  }
+
+  confirmPermanentLocation(confirm) {
+    const { permanentLocation, prevPermanentLocation } = this.state;
+    const confirmPermanentLocation = false;
+    const value = (confirm) ? permanentLocation.id : prevPermanentLocation.id;
+    const prevPermanentLoc = (confirm) ? permanentLocation : prevPermanentLocation;
     setTimeout(() => this.props.change('permanentLocationId', value));
-    this.setState({ confirmLocation, prevLocation: prevLoc });
+    this.setState({ confirmPermanentLocation, prevPermanentLocation: prevPermanentLoc });
+  }
+
+  confirmTemporaryLocation(confirm) {
+    const { temporaryLocation, prevTemporaryLocation } = this.state;
+    const confirmTemporaryLocation = false;
+    const value = (confirm) ? temporaryLocation.id : prevTemporaryLocation.id;
+    const prevTemporaryLoc = (confirm) ? temporaryLocation : prevTemporaryLocation;
+    setTimeout(() => this.props.change('temporaryLocationId', value));
+    this.setState({ confirmTemporaryLocation, prevTemporaryLocation: prevTemporaryLoc });
   }
 
   render() {
@@ -85,11 +130,11 @@ class HoldingsForm extends React.Component {
       copy,
     } = this.props;
     const formatMsg = this.props.formatMsg;
-    const { confirmLocation } = this.state;
+    const { confirmPermanentLocation, confirmTemporaryLocation } = this.state;
 
     /* Menus for Add Item workflow */
-    const addHoldingsLastMenu = <PaneMenu><Button buttonStyle="primary paneHeaderNewButton" id="clickable-create-item" type="submit" title={formatMsg({ id: 'ui-inventory.createHoldingsRecord' })} disabled={(pristine || submitting) && !copy} onClick={handleSubmit}>Create holdings record</Button></PaneMenu>;
-    const editHoldingsLastMenu = <PaneMenu><Button buttonStyle="primary paneHeaderNewButton" id="clickable-update-item" type="submit" title={formatMsg({ id: 'ui-inventory.updateHoldingsRecord' })} disabled={(pristine || submitting) && !copy} onClick={handleSubmit}>Update holdings record</Button></PaneMenu>;
+    const addHoldingsLastMenu = <PaneMenu><Button buttonStyle="primary paneHeaderNewButton" id="clickable-create-item" type="submit" title={formatMsg({ id: 'ui-inventory.createHoldingsRecord' })} disabled={(pristine || submitting) && !copy} onClick={handleSubmit(this.onSave)}>Create holdings record</Button></PaneMenu>;
+    const editHoldingsLastMenu = <PaneMenu><Button buttonStyle="primary paneHeaderNewButton" id="clickable-update-item" type="submit" title={formatMsg({ id: 'ui-inventory.updateHoldingsRecord' })} disabled={(pristine || submitting) && !copy} onClick={handleSubmit(this.onSave)}>Update holdings record</Button></PaneMenu>;
 
     const platformOptions = (referenceTables.platforms || []).map(l => ({
       label: l.name,
@@ -131,9 +176,24 @@ class HoldingsForm extends React.Component {
                   component={LocationSelection}
                   fullWidth
                   marginBottom0
-                  onChange={loc => this.selectLocation(loc)}
+                  onSelect={loc => this.selectPermanentLocation(loc)}
                 />
-                <LocationLookup onLocationSelected={loc => this.selectLocation(loc)} />
+                <LocationLookup onLocationSelected={loc => this.selectPermanentLocation(loc)} />
+              </Col>
+            </Row>
+            <Row >
+              <Col sm={5} smOffset={1}>
+                <Field
+                  label={formatMsg({ id: 'ui-inventory.temporaryLocation' })}
+                  placeholder={formatMsg({ id: 'ui-inventory.selectTemporaryLocation' })}
+                  name="temporaryLocationId"
+                  id="additem_temporarylocation"
+                  component={LocationSelection}
+                  fullWidth
+                  marginBottom0
+                  onSelect={loc => this.selectTemporaryLocation(loc)}
+                />
+                <LocationLookup onLocationSelected={loc => this.selectTemporaryLocation(loc)} />
               </Col>
             </Row>
             <Row>
@@ -162,14 +222,26 @@ class HoldingsForm extends React.Component {
             </Row>
             <FieldArray name="holdingsStatements" component={renderStatements} formatMsg={formatMsg} />
             <ConfirmationModal
-              open={confirmLocation}
+              id="confirmPermanentLocationModal"
+              open={confirmPermanentLocation}
               heading={formatMsg({ id: 'ui-inventory.confirmLocation.header' })}
               message={formatMsg({ id: 'ui-inventory.confirmLocation.message' })}
               confirmLabel={formatMsg({ id: 'ui-inventory.confirmLocation.selectBtn' })}
               buttonStyle="default"
               cancelButtonStyle="primary"
-              onConfirm={() => { this.confirmLocation(true); }}
-              onCancel={() => { this.confirmLocation(false); }}
+              onConfirm={() => { this.confirmPermanentLocation(true); }}
+              onCancel={() => { this.confirmPermanentLocation(false); }}
+            />
+            <ConfirmationModal
+              id="confirmTemporaryLocationModal"
+              open={confirmTemporaryLocation}
+              heading={formatMsg({ id: 'ui-inventory.confirmLocation.header' })}
+              message={formatMsg({ id: 'ui-inventory.confirmLocation.message' })}
+              confirmLabel={formatMsg({ id: 'ui-inventory.confirmLocation.selectBtn' })}
+              buttonStyle="default"
+              cancelButtonStyle="primary"
+              onConfirm={() => { this.confirmTemporaryLocation(true); }}
+              onCancel={() => { this.confirmTemporaryLocation(false); }}
             />
           </Pane>
         </Paneset>