+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 * Add item form, holdings form (UIIN-5, UIIN-20)
 * Add list of holdings records and items under instance view (UIIN-15,16,29,31,33)
 * Project renamed from ui-instances to ui-inventory. UIIN-17.
-* Update Okapi dependencies, inventory: 5.0, instance-storage 4.2, item-storage: 5.0
+* Update Okapi dependencies, inventory: 5.0 6.0, instance-storage 4.2, item-storage: 5.0
 * Add Okapi dependencies on reference tables
 * Use PropTypes, not React.PropTypes. STRIPES-427.
 * Validation. Fixes UIIN-19.
@@ -77,6 +77,12 @@
 * Fix uniqueness validation of Item bar-code. UIIN-189.
 * Update to current users interface. Refs UIU-495.
 * Update import paths for some stripes-components. Refs STCOM-277.
+* Refactor to avoid deprecated React lifecycle methods.
+* Add temporary location to HoldingsRecord form and details view. UIIN-194.
+* Add item's permanent location to Item form and details view. UIIN-195.
+* Display effective location on Item details view. UIIN-196.
+* Enable un-setting a location. Fixes UIIN-198.
+* Require inventory 5.3, item-storage 5.3, holdings-storage 1.3. UIIN-194,-195,-196
 
 ## [1.0.0](https://github.com/folio-org/ui-instances/tree/v1.0.0) (2017-09-08)
 [Full Changelog](https://github.com/folio-org/ui-items/compare/v0.0.1...v1.0.0)